+++ playground/socket/index.ts
@@ -1,12 +1,11 @@
 import { listen } from 'socket.io';
 import {
   Middleware,
-  GlobalMiddleware,
   ServerMiddleware,
   Event,
   Args,
   attachControllers,
-  Namespace,
+  Controller,
   Socket
 } from '@decorators/socket';
 
@@ -14,9 +13,8 @@ const server = listen(3000);
 
 class SocketWrapper {
   constructor(private socket: SocketIO.Socket) {}
-
-  log() {
-    console.log('Log');
+  log(message: string) {
+    console.log('Log::', message);
   }
 }
 
@@ -24,32 +22,33 @@ class SocketWrapper {
   console.log('Global Server Middleware');
   next();
 })
-@GlobalMiddleware((io, socket, packet, next) => {
-  console.log('Global socket middleware');
-  next();
-})
 @Middleware((io, socket, packet, next) => {
-  console.log('Controller based middleware');
+  console.log('Global Socket Middleware');
   next();
 })
-@Namespace('/messaging')
+@Controller('/messaging', (io, socket, packet, next) => {
+  console.log('Controller Middleware');
+  next();
+})
 class FirstController {
 
+  deps: any[];
+
   constructor(...args) {
-    console.log(args);
+    console.log('Controller Instantiation', args);
+    this.deps = args;
   }
 
   @Event('message', (io, socket, packet, next) => {
-    console.log('Message middleware');
+    console.log('Event Middleware');
     next();
   })
   onMessage(@Args() message, @Socket(SocketWrapper) socket: SocketWrapper) {
-    socket.log();
-    console.log(`Message:  ${message}`);
+    socket.log(`Message:  ${message} ${this.deps.toString()}`);
   }
 
 }
 
 attachControllers(server, [
-  { provide: FirstController, deps: [1, 2, 3]}
+  { provide: FirstController, deps: [1, 2, 3] }
 ]);