+++ socket/src/decorators/index.ts
@@ -1,4 +1,4 @@
-export { Namespace } from './ns';
-export { ServerMiddleware, GlobalMiddleware, Middleware } from './middleware';
+export { Controller } from './controller';
+export { ServerMiddleware, Middleware } from './middleware';
 export { Connection, Disconnect, GlobalEvent, Event } from './events';
 export { IO, Socket, Args, Callback } from './params';