+++ socket/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Socket#2.0.0
+* Renamed **GlobalMiddleware** for socket global middleware into **Middleware**
+* Renamed **Namespace** to **@Controller(namespace: string, middleware?: Function | Function[])**
+* Removed **@Middleware(middleware: Function | Function[])** - use Controller based middleware
 * Removed deprecated methods
   * **bootstrapSocketIO(io: SocketIO.Server, Controllers)** -  Attaches controllers to IO server - use **attachControllers** instead
   * **attachControllerToSocket(io: SocketIO.Server, socket: SocketIO.Socket, Controllers)** -  Attaches controllers to Socket - use **attachControllersToSocket** instead
@@ -76,7 +79,7 @@ class SocketWrapper {
   * **bootstrapSocketIO(rootController: Controller)**
   ##### Class
   * **@Connect(serverOrPort: number | string | HttpServer, opts?: any)**
-  * **@Middleware(fn: Function)** 
+  * **@Middleware(fn: Function)**
   ##### Method
   * **@OnIO(event: string)**
   * **@OnConnect()**